Adding updated code, plot

+++ Lab_3/sensor_read.py
@@ -24,9 +24,9 @@ while True:
     if len(data) == 6:
         time.append(float(data[0]))
         left_sensor.append(float(data[1]))
-        right_sensor.append(int(data[2]))
-        left_motor.append(int(data[3]))
-        right_motor.append(int(data[4]))
+        right_sensor.append(float(data[2]))
+        left_motor.append(float(data[3]))
+        right_motor.append(float(data[4]))
         if(int(data[5]) == 1):
             break
